requirements.txt
Django==4.1.4
djangorestframework==3.14.0
django-webpack-loader==0.7.0
django-cors-headers==3.13.0
dj-database-url==1.2.0
requests==2.28.1
requests-mock==1.10.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.3.0
python-dotenv==0.21.0